Don't send OK when the channel is not ready
In Cordova-WP8 3.7.0 there was a problem that NO_RESULT was treated as error (CB-8580), but this was fixed over a year ago. I'm guessing the current implementation can lead to people having trouble registering (receiving a push token) because of the OK we're sending when the channel is not ready yet. Let's try and use NO_RESULT again (which should work fine with Cordova-WP8 3.8.0 and up).